requirements.txt
pandas
matplotlib
requests
python-dotenv
holidays
openpyxl
pyrogram
tgcrypto

# аналитика
scikit-learn
tqdm

# карты и геоданные
plotly
geopandas
shapely
pyproj
fiona
pyarrow
kaleido




# логирование (опционально)
loguru